fix android migration
