+++ examples/react-native-todo-list/metro.config.js
@@ -1,20 +1,7 @@
-// const { getDefaultConfig } = require("metro-config");
+const { getDefaultConfig, mergeConfig } = require("@react-native/metro-config");
 
-// module.exports = (async () => {
-//   const {
-//     resolver: { sourceExts, assetExts },
-//   } = await getDefaultConfig();
-//   return {
-//     transformer: {
-//       babelTransformerPath: require.resolve("react-native-svg-transformer"),
-//     },
-//     resolver: {
-//       assetExts: assetExts.filter((ext) => ext !== "svg"),
-//       sourceExts: [...sourceExts, "svg"],
-//     },
-//   };
-// })();
-const {getDefaultConfig, mergeConfig} = require('@react-native/metro-config');
+const defaultConfig = getDefaultConfig(__dirname);
+const { assetExts, sourceExts } = defaultConfig.resolver;
 
 /**
  * Metro configuration
@@ -22,6 +9,14 @@ const {getDefaultConfig, mergeConfig} = require('@react-native/metro-config');
  *
  * @type {import('metro-config').MetroConfig}
  */
-const config = {};
+const config = {
+  transformer: {
+    babelTransformerPath: require.resolve('react-native-svg-transformer'),
+  },
+  resolver: {
+    assetExts: assetExts.filter((ext) => ext !== 'svg'),
+    sourceExts: [...sourceExts, 'svg'],
+  },
+};
 
-module.exports = mergeConfig(getDefaultConfig(__dirname), config);
+module.exports = mergeConfig(defaultConfig, config);
