+++ examples/react-native-todo-list/package.json
@@ -17,12 +17,10 @@
     "base-64": "^1.0.0",
     "react": "18.2.0",
     "react-native": "^0.73.0",
-    "react-native-svg": "^14.0.0",
-    "react-native-svg-transformer": "^1.0.0"
+    "react-native-svg": "^15.1.0",
+    "react-native-svg-transformer": "^1.3.0"
   },
   "devDependencies": {
-    "@types/base-64": "^1.0.0",
-    "@types/react-native": "^0.69.1",
     "@babel/core": "^7.20.0",
     "@babel/preset-env": "^7.20.0",
     "@babel/runtime": "^7.20.0",
@@ -30,7 +28,9 @@
     "@react-native/eslint-config": "^0.73.1",
     "@react-native/metro-config": "^0.73.0",
     "@react-native/typescript-config": "^0.73.1",
+    "@types/base-64": "^1.0.0",
     "@types/react": "^18.2.6",
+    "@types/react-native": "^0.73.0",
     "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.6.3",
     "eslint": "^8.19.0",
